add fastapi instrumentation metrics
