+++ pyproject.toml
@@ -34,6 +34,7 @@ dependencies = [
     "jinja2>=3.1.2",
     "dependency-injector>=4.41.0",
     "grpcio>=1.59.0",
+    "prometheus-fastapi-instrumentator>=6.1.0",
 ]
 requires-python = ">=3.11"
 license = {text = "MIT"}
